test: Update the test report generator to enhance module categorization and report generation functionality
1. Added ignores for `__pycache__` and `*.pyc` files in `.gitignore`.
2. Refactored `generate-report.py`, improving the description and parameter explanations of the main function, replacing Chinese with English.
3. Added new logging functionality to `run-ut.sh`, supporting execution from a specific step and providing usage instructions.
4. Introduced a new `report_generator` package containing the core logic and parsers for test report generation, improving code structure clarity and maintainability.

Log: These changes make test report generation more flexible and user-friendly while enhancing code readability and modularity.
