+++ .gitignore
@@ -93,3 +93,7 @@ dde-file-manager-lib/test_shutil/property/oneProperty.pro
 .cursorindexingignore
 .cursor
 .specstory
+
+# autotests
+__pycache__
+*.pyc
