+++ autotests/report_generator/__init__.py
@@ -1,0 +1,17 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+
+# SPDX-FileCopyrightText: 2025 UnionTech Software Technology Co., Ltd.
+#
+# SPDX-License-Identifier: GPL-3.0-or-later
+
+"""
+DDE File Manager Test Report Generator Package
+"""
+
+__version__ = "1.0.0"
+__author__ = "DDE File Manager Team"
+
+from .main import TestReportGenerator
+
+__all__ = ['TestReportGenerator']
